// src/enums/mutationTypes.ts
export enum MutationTypes {
  SET_LOGIN = "SET_LOGIN",
  SET_USER_INFO = "SET_USER_INFO",
  SET_AVATAR_PATH = "SET_AVATAR_PATH",
  SET_ANNOUNCEMENT_LIST = "SET_ANNOUNCEMENT_LIST",
  SET_BANNER_LIST = "SET_BANNER_LIST",
  SET_LOCATION = "SET_LOCATION",
  SET_COUNT = "SET_COUNT",
  SET_MY_MISSIONS = "SET_MY_MISSIONS",
  SET_MY_UNCHECKED_MISSIONS = "SET_MY_UNCHECKED_MISSIONS",
  SET_MY_ALL_MISSIONS = "SET_MY_ALL_MISSIONS",
  SET_CURRENT_MISSION = "SET_CURRENT_MISSION",
  SET_CURRENT_MISSION_INFO = "SET_CURRENT_MISSION_INFO",
  SET_CURRENT_MISSION_MEMBERS = "SET_CURRENT_MISSION_MEMBERS",
  SET_FACE_RECOGNITION_HISTORY = "SET_FACE_RECOGNITION_HISTORY",
  UPDATE_MISSION_VOLUNTEER_LOCATION = "UPDATE_MISSION_VOLUNTEER_LOCATION",
  UPDATE_MISSION_VOLUNTEER_OFFLINE = "UPDATE_MISSION_VOLUNTEER_OFFLINE",
  SET_WS = "SET_WS",
  SET_TIM_SDK_READY = "SET_TIM_SDK_READY",
  SET_TIM_MY_INFO = "SET_TIM_MY_INFO",
  SET_TIM_MY_GROUPS = "SET_TIM_MY_GROUPS",
  SET_TIM_ALL_CONVERSATION = "SET_TIM_ALL_CONVERSATION",
  RESET_CURRENT_CONVERSATION = "RESET_CURRENT_CONVERSATION",
  UPDATE_CURRENT_CONTERSATION = "UPDATE_CURRENT_CONTERSATION",
  UNSHIFT_MESSAGE_LIST = "UNSHIFT_MESSAGE_LIST",
  RECEIVE_MESSAGE = "RECEIVE_MESSAGE",
  SEND_MESSAGE = "SEND_MESSAGE",
  ADD_UNREAD_MESSAGE_NUMBER = "ADD_UNREAD_MESSAGE_NUMBER",
  CLEAR_UNREAD_MESSAGE_NUMBER = "CLEAR_UNREAD_MESSAGE_NUMBER",
  SET_MY_REPAIRORDERS = "SET_MY_REPAIRORDERS",
  SET_WORKER_REPAIRORDERS = "SET_WORKER_REPAIRORDERS",//获取全部师傅订单
  SET_WORKER_WORKING_REPAIRORDERS = "SET_WORKER_WORKING_REPAIRORDERS",//获取进行中的订单
  SET_WORKER_UNCONFIRMED_REPAIRORDERS = "SET_WORKER_UNCONFIRMED_REPAIRORDERS",//获取待确认的订单
  SET_WORKER_FINISHED_REPAIRORDERS = "SET_WORKER_FINISHED_REPAIRORDERS",//获取已完成的订单
  SET_WORKER_BACK_REPAIRORDERS = "SET_WORKER_BACK_REPAIRORDERS",//获取已退单的订单
}